Simplify onboarding flow. Remove root viewmodel, onboardingviewmodel. Move logic to parent viewmodel. Add close button to onboarding screen. Rename Onboarding folder -> OnBoarding
