+++ Sources/Domain/Model/Presentation/OnBoardingScreen.swift
@@ -1,0 +1,15 @@
+//
+//  OnBoardingScreen.swift
+//  Tumble
+//
+//  Created by Adis Veletanlic on 2023-05-27.
+//
+
+import Foundation
+
+struct BoardingScreen: Identifiable {
+    var id = UUID().uuidString
+    var image: String
+    var title: String
+    var description: String
+}
