+++ Sources/TumbleApp.swift
@@ -12,18 +12,17 @@ struct TumbleApp: App {
     @UIApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
     
     init() {
-        // Cache the environment variable in app storage (user defaults).
+        /// Cache the environment variable in app storage (user defaults).
         if let networkSettings = ProcessInfo.processInfo.environment["NETWORK_SETTINGS"] {
             UserDefaults.standard.set(networkSettings, forKey: StoreKey.networkSettings.rawValue)
         }
-        // Initialize dependency providers
+        /// Initialize dependency providers
         _ = Dependencies()
     }
     
     var body: some Scene {
         WindowGroup {
-            Root(viewModel: ViewModelFactory().makeViewModelRoot())
-                .onAppear()
+            Root()
         }
     }
 }
